Added perror() call to rs.c

+++ md5.h
@@ -135,7 +135,7 @@ extern void *md5_read_ctx __P ((const struct md5_ctx *ctx, void *resbuf));
 /* Compute MD5 message digest for bytes read from STREAM.  The
    resulting message digest number will be written into the 16 bytes
    beginning at RESBLOCK.  */
-extern int md5_stream __P ((FILE *stream, void *resblock));
+extern ssize_t md5_stream __P ((FILE *stream, void *resblock));
 
 /* Compute MD5 message digest for LEN bytes beginning at BUFFER.  The
    result is always in little endian byte order, so that a byte-wise
